fix: add event polling for hvac on and off, chore: consolidate remote result functions
* add poll and unlock to climate buttons
* add 019 and 021 to action types polled (climate and EV)
* consolidate remote poll and tack functions into one poll and unlock
* move button sleep interval to const.py
