+++ custom_components/mazda_cs/climate.py
@@ -105,6 +105,7 @@ class MazdaClimateEntity(MazdaEntity, ClimateEntity):
                 self._attr_min_temp = 15.5
                 self._attr_max_temp = 28.5
 
+        self._command_in_progress = False
         self._update_state_attributes()
 
     @callback
@@ -150,14 +151,19 @@ class MazdaClimateEntity(MazdaEntity, ClimateEntity):
 
     async def async_set_hvac_mode(self, hvac_mode: HVACMode) -> None:
         """Set a new HVAC mode."""
+        if self._command_in_progress:
+            return
         command_utc = datetime.now(timezone.utc)
         if hvac_mode == HVACMode.HEAT_COOL:
             await self.client.turn_on_hvac(self.vehicle_id)
-            self._track_remote_result("hvacOn", command_utc)
+            action = "hvacOn"
         elif hvac_mode == HVACMode.OFF:
             await self.client.turn_off_hvac(self.vehicle_id)
-            self._track_remote_result("hvacOff", command_utc)
-
+            action = "hvacOff"
+        else:
+            return
+        self._command_in_progress = True
+        self.hass.async_create_task(self._poll_and_unlock(action, command_utc))
         self._handle_coordinator_update()
 
     async def async_set_temperature(self, **kwargs: Any) -> None:
