+++ custom_components/mazda_cs/switch.py
@@ -193,6 +193,7 @@ class MazdaChargingSwitch(MazdaEntity, SwitchEntity):
         super().__init__(client, coordinator, index)
 
         self._attr_unique_id = self.vin
+        self._command_in_progress = False
 
     @property
     def is_on(self):
@@ -209,14 +210,20 @@ class MazdaChargingSwitch(MazdaEntity, SwitchEntity):
 
     async def async_turn_on(self, **kwargs: Any) -> None:
         """Start charging the vehicle."""
+        if self._command_in_progress:
+            return
         command_utc = datetime.now(timezone.utc)
         await self.client.start_charging(self.vehicle_id)
-        self._track_remote_result("chargeStart", command_utc)
+        self._command_in_progress = True
+        self.hass.async_create_task(self._poll_and_unlock("chargeStart", command_utc))
         await self.refresh_status_and_write_state()
 
     async def async_turn_off(self, **kwargs: Any) -> None:
         """Stop charging the vehicle."""
+        if self._command_in_progress:
+            return
         command_utc = datetime.now(timezone.utc)
         await self.client.stop_charging(self.vehicle_id)
-        self._track_remote_result("chargeStop", command_utc)
+        self._command_in_progress = True
+        self.hass.async_create_task(self._poll_and_unlock("chargeStop", command_utc))
         await self.refresh_status_and_write_state()
